This is the baseline I've created so far.
The bullet file will probably leave at some point, and the quitter file might be merged into another file to manage all system operations. But that's for the future, and this is now. There is also some random stuff that should be cleaned up, and the next commit should do that, as well as add comments where they are necessary.